requirements.txt
requests~=2.32.3
flask~=3.1.0
python-dotenv